# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Refactoring",
        "title": "online-judge",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Eutop1a",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:65533",
    "basePath": "/api/v1",
    "paths": {
        "/admin/problem/create": {
            "post": {
                "description": "创建新题目接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json",
                    "multipart/form-data"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "创建新题目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "题目标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "题目内容",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "题目难度",
                        "name": "difficulty",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "时间限制",
                        "name": "max_runtime",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "内存限制",
                        "name": "max_memory",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "测试样例集",
                        "name": "test_cases",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            }
        },
        "/admin/problem/{problem_id}": {
            "put": {
                "description": "更新题目信息接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "更新题目信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "题目ID",
                        "name": "problem_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "题目标题",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "题目内容",
                        "name": "content",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "题目难度",
                        "name": "difficulty",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "时间限制",
                        "name": "max_runtime",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "内存限制",
                        "name": "max_memory",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "测试样例集",
                        "name": "test_cases",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除题目接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "删除题目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "题目ID",
                        "name": "problem_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            }
        },
        "/admin/users/add-admin": {
            "post": {
                "description": "添加管理员接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "添加管理员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            }
        },
        "/admin/users/{user_id}": {
            "delete": {
                "description": "删除用户接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            }
        },
...
        "/problem/id": {
            "post": {
                "description": "获取题目ID接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Problem API"
                ],
                "summary": "获取题目ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "题目title不存在",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            }
        },
        "/problem/list": {
            "get": {
                "description": "获取题目列表接口",
                "tags": [
                    "Problem API"
                ],
                "summary": "获取题目列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            }
        },
        "/problem/{problem_id}": {
            "get": {
                "description": "获取单个题目详细接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json",
                    "multipart/form-data"
                ],
                "tags": [
                    "Problem API"
                ],
                "summary": "获取单个题目详细",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "题目ID",
                        "name": "problem_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller._Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "description": "omitempty 字段为空就忽略"
                },
                "msg": {}
            }
        }
    }
}